Adds Capybara test for subscribing
Capybara-webkit is being used to test the Stripe checkout process. It's a bit brittle, and doesn't seem to lend well towards using something like puffing-billy for request mocking. I get the feeling that this will be the norm for several feature specs, so I've excluded the features folder from `rake check`, which avoids having to set it up for CI testing.